fix: 将寄存器映射的键类型从int改为int64_t
避免潜在的整数溢出问题，确保寄存器可以处理更大的键值

+++ Interpreter.hpp
@@ -12,7 +12,7 @@ public:
 
   int pc=0;
   void add(std::string str);
-  std::unordered_map<int,int64_t> reg;
+  std::unordered_map<int64_t,int64_t> reg;
   void debug();
   void run();//不建议使用
   void step();
